# Add some loading / LazyArtifacts precompiles to the sysimage